Split auth tests into multiple files to make use of Jest per-file DOM sandbox
